Rename: 폴더 구조 변경 및 인덱스 정리
