+++ src/components/rootPage/Board/Board.jsx
@@ -1,8 +1,9 @@
 import { Suspense, useEffect, useState } from 'react';
 import { Container } from '@mantine/core';
-import { CategoryPicker, Posters, Carousels } from '.';
-import { PosterSkeleton } from './Posters';
-import { goToTop } from '../../utils';
+import { PosterSkeleton } from '../Posters';
+import { Posters, Carousels, CategoryPicker } from '..';
+
+import { goToTop } from '../../../utils';
 
 const Board = () => {
   const [mediaType, setMediaType] = useState('movie');
