+++ src/components/rootPage/index.js
@@ -1,12 +1,6 @@
+export { default as Board } from './Board/Board';
+export { default as TopButton } from './TopButton/TopButton';
 export { Banner } from './Banner';
+export { Posters } from './Posters';
 export { CategoryPicker } from './CategoryPicker';
 export { Carousels } from './Carousels';
-
-export { default as Board } from './Board';
-export { Posters } from './Posters';
-export { default as Poster } from './Posters/Poster';
-
-export { default as Slide } from './Slide';
-
-export { default as MoreButton } from './Carousels/CarouselSkeleton';
-export { default as TopButton } from './TopButton';
